Fixed entities spinning
This is a vanilla issue and I cannot for the life of me figure out why mojang have decided to have this happen. Iv'e explained the issue to Neus, but heres a basic rundown. The width for the navigators to search to determin if a path node is compleated is as follows:

`this.entity.width > 0.75F ? this.entity.width / 2.0F : 0.75F - this.entity.width / 2.0F`. It doesn't take a genius to see that if the width is larger than 0.75 then it's going to start decresasing the search distance. An entity at width 1 (what these entities originally were), they'd have a path node range of `0.25`, resulting in the oft not being able to detect they're at path node and therefore cannot mark it as compelated. They would then spin around at the path node point until they either give up or the path point goes within range.